Assignment 2/requirements.txt
absl-py==0.2.0
astor==0.6.2
attrs==17.4.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
cliff==2.11.0
cmd2==0.8.5
configparser==3.5.0
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
futures==3.1.1
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==1.0.1
idna==2.6
imbalanced-learn==0.3.3
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.3.3
kaggle-cli==0.12.13
Keras==2.1.6
kiwisolver==1.0.1
livelossplot==0.1.7
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
MechanicalSoup==0.10.0
mistune==0.8.3
more-itertools==4.1.0
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nose==1.3.7
notebook==5.4.1
numpy==1.14.3
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.2.0
pbr==4.0.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prettytable==0.7.2
progressbar2==3.37.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
pyparsing==2.2.0
pyperclip==1.6.0
pytest==3.5.1
python-dateutil==2.7.2
python-utils==2.3.0
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
stevedore==1.28.0
sympy==1.1.1
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
tqdm==4.23.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
